refactor: große Komponenten aufbrechen — 12 Hooks, 4 Sub-Components (#212)
5 Komponenten von >700 auf <500 Zeilen reduziert:
- SessionTemplateEditor: 973 → 338
- Upload: 734 → 481
- WeeklyPlan: 740 → 395
- TrainingPlanEditor: 718 → 459
- PhaseWeeklyTemplateEditor: 788 → 268

12 neue Custom Hooks extrahiert:
useExerciseListEditor, useExerciseSuggestions, useSessionTemplateForm,
useUploadForm, usePlannedSessionLinking, useStrengthUpload,
useWeeklyPlan, useWeeklyPlanDragDrop, usePlanForm, usePlanSubmit,
usePhaseWeeklyTemplate, weeklyPlanUtils

4 Sub-Components ausgelagert:
StrengthExercisesEditSection, StrengthExercisesReadView,
RunningReadView, TemplateSessionEditor

TSC, ESLint (0 warnings), Prettier, Vitest (145 tests) grün.
