+++ frontend/src/utils/weeklyPlanUtils.ts
@@ -1,0 +1,38 @@
+/**
+ * Utility functions for weekly plan date handling and formatting.
+ */
+import { CATEGORY_LABELS } from '@/constants/training';
+
+export function categoryLabel(key: string): string {
+  return CATEGORY_LABELS[key] ?? key;
+}
+
+export function getMondayOfWeek(d: Date): string {
+  const copy = new Date(d);
+  const day = copy.getDay();
+  const diff = day === 0 ? -6 : 1 - day;
+  copy.setDate(copy.getDate() + diff);
+  return copy.toISOString().split('T')[0];
+}
+
+export function addWeeks(dateStr: string, weeks: number): string {
+  const d = new Date(dateStr);
+  d.setDate(d.getDate() + weeks * 7);
+  return d.toISOString().split('T')[0];
+}
+
+export function formatDateRange(weekStart: string): string {
+  const start = new Date(weekStart);
+  const end = new Date(weekStart);
+  end.setDate(end.getDate() + 6);
+
+  const startDay = start.getDate();
+  const endDay = end.getDate();
+  const startMonth = start.toLocaleDateString('de-DE', { month: 'short' });
+  const endMonth = end.toLocaleDateString('de-DE', { month: 'short' });
+
+  if (startMonth === endMonth) {
+    return `${startDay}. – ${endDay}. ${startMonth}`;
+  }
+  return `${startDay}. ${startMonth} – ${endDay}. ${endMonth}`;
+}
